requirements.txt
absl-py==0.1.10
appnope==0.1.0
automium==0.2.6
automium-web==0.1.1
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.1.2
Django==1.11.5
Flask==0.12.2
futures==3.1.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
Keras==2.1.3
learn==1.0.0
legacy==0.1.6
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
networkx==2.0
numpy==1.14.0
olefile==0.44
opencv-contrib-python==3.4.0.12
opencv-python==3.3.0.10
parso==0.1.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-resize-image==1.1.11
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.0
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
stevedore==1.27.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
wcwidth==0.1.7
Werkzeug==0.14.1
